♻️: shared/ui Input 컴포넌트 리팩토링 - 분리
