+++ src/shared/ui/index.tsx
@@ -1,27 +1,8 @@
 import * as React from 'react';
-import { ComponentPropsWithRef, forwardRef, HTMLAttributes } from 'react';
+import { forwardRef, HTMLAttributes } from 'react';
 import * as SelectPrimitive from '@radix-ui/react-select';
 import * as DialogPrimitive from '@radix-ui/react-dialog';
 import { Check, ChevronDown, X } from 'lucide-react';
-
-type InputProps = ComponentPropsWithRef<'input'> & {
-  // 추가적인 커스텀 props가 필요하다면 여기에 정의
-};
-
-// 입력 컴포넌트
-export const Input = forwardRef<HTMLInputElement, InputProps>(
-  ({ className, type, ...props }, ref) => {
-    return (
-      <input
-        type={type}
-        className={`flex h-10 w-full rounded-md border border-input bg-white px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 ${className}`}
-        ref={ref}
-        {...props}
-      />
-    );
-  },
-);
-Input.displayName = 'Input';
 
 // 기본이 되는 타입 정의
 type BaseProps = HTMLAttributes<HTMLElement> & {
